Amazon
To use this script, you'll need to first download your order report from Amazon by following the steps mentioned in the comments of the code. Once you have the report file downloaded, replace the REPORT_FILE_PATH variable with the path to your report file, and run the script. It will read the CSV file and calculate the total amount spent on all orders in the report.